bump up the Google group mention
currently joining the Google group is not mentioned until the
end of the index page. Often times people send a message to
hello@cugos.org just to say they'd like to join the group. thought
it would be nice to explicity mention that if they're interested in
joining they should joince the Google group and then say hi at
hello@cugos.org.